requirements.txt
requests>=2.31.0
python-whois>=0.9.4
scapy>=2.5.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
lxml>=4.9.0
urllib3>=2.0.0
